Rename utils module to avoid import conflicts
