+++ rate_utils.py
